add HandlerTests.GetWeekSummaries.cs
